+++ samples/sampler/stories/default/CheckboxItem.js
@@ -1,12 +1,12 @@
 import {action} from '@enact/storybook-utils/addons/actions';
 import {boolean, select, text} from '@enact/storybook-utils/addons/knobs';
 import {mergeComponentMetadata, nullify} from '@enact/storybook-utils';
-import UiToggleItem, {ToggleItemBase as UiToggleItemBase} from '@enact/ui/ToggleItem';
 import UiItem from '@enact/ui/Item';
 import {listIcons} from './icons';
 import {storiesOf} from '@storybook/react';
 
 import CheckboxItem from '@enact/moonstone/CheckboxItem';
+import UiToggleItem, {ToggleItemBase as UiToggleItemBase} from '@enact/moonstone/internal/ToggleItem';
 import ToggleItem from '@enact/moonstone/ToggleItem';
 import Item, {ItemBase} from '@enact/moonstone/Item';
 import Icon from '@enact/moonstone/Icon';