+++ samples/sampler/stories/default/SelectableItem.js
@@ -1,10 +1,10 @@
 import {action} from '@enact/storybook-utils/addons/actions';
 import {boolean, select, text} from '@enact/storybook-utils/addons/knobs';
 import {mergeComponentMetadata, nullify} from '@enact/storybook-utils';
-import UiToggleItem, {ToggleItemBase as UiToggleItemBase} from '@enact/ui/ToggleItem';
 import {storiesOf} from '@storybook/react';
 
 import Icon from '@enact/moonstone/Icon';
+import UiToggleItem, {ToggleItemBase as UiToggleItemBase} from '@enact/moonstone/internal/ToggleItem';
 import Item, {ItemBase} from '@enact/moonstone/Item';
 import SelectableItem from '@enact/moonstone/SelectableItem';
 import ToggleItem from '@enact/moonstone/ToggleItem';