+++ samples/sampler/stories/default/VirtualGridList.js
@@ -3,9 +3,9 @@ import {boolean, number, select} from '@enact/storybook-utils/addons/knobs';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import ri from '@enact/ui/resolution';
 import {storiesOf} from '@storybook/react';
-import {VirtualListBase as UiVirtualListBase} from '@enact/ui/VirtualList';
 
 import {GridListImageItem} from '@enact/moonstone/GridListImageItem';
+import {VirtualListBase as UiVirtualListBase} from '@enact/moonstone/internal/VirtualList';
 import {ScrollableBase as UiScrollableBase} from '@enact/moonstone/internal/Scrollable';
 import {VirtualGridList, VirtualListBase} from '@enact/moonstone/VirtualList';