+++ samples/sampler/stories/default/VirtualList.js
@@ -2,10 +2,10 @@ import {action} from '@enact/storybook-utils/addons/actions';
 import {boolean, number, select} from '@enact/storybook-utils/addons/knobs';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import ri from '@enact/ui/resolution';
-import {ScrollableBase as UiScrollableBase} from '@enact/moonstone/internal/Scrollable';
 import {storiesOf} from '@storybook/react';
-import {VirtualListBase as UiVirtualListBase} from '@enact/ui/VirtualList';
 
+import {ScrollableBase as UiScrollableBase} from '@enact/moonstone/internal/Scrollable';
+import {VirtualListBase as UiVirtualListBase} from '@enact/moonstone/internal/VirtualList';
 import Item from '@enact/moonstone/Item';
 import VirtualList, {VirtualListBase} from '@enact/moonstone/VirtualList';