+++ samples/sampler/stories/qa/VirtualGridList.js
@@ -9,7 +9,7 @@ import {boolean, number, select} from '@enact/storybook-utils/addons/knobs';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import ri from '@enact/ui/resolution';
 
-import {VirtualListBase as UiVirtualListBase} from '@enact/ui/VirtualList/VirtualListBase';
+import {VirtualListBase as UiVirtualListBase} from '@enact/moonstone/internal/VirtualList';
 import {storiesOf} from '@storybook/react';
 import PropTypes from 'prop-types';
 import {useCallback, useEffect, useRef, useState} from 'react';