+++ samples/sampler/stories/qa/VirtualList.js
@@ -10,7 +10,7 @@ import {action} from '@enact/storybook-utils/addons/actions';
 import {boolean, number, select} from '@enact/storybook-utils/addons/knobs';
 import {mergeComponentMetadata} from '@enact/storybook-utils';
 import ri from '@enact/ui/resolution';
-import {VirtualListBase as UiVirtualListBase} from '@enact/ui/VirtualList';
+import {VirtualListBase as UiVirtualListBase} from '@enact/moonstone/internal/VirtualList';
 import {storiesOf} from '@storybook/react';
 import PropTypes from 'prop-types';
 import {useCallback, useEffect, useRef, useState} from 'react';