+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/internal/RadioDecorator` component, transferred from `ui/RadioDecorator`
 - `moonstone/internal/Scrollable` component, transferred from `ui/Scrollable`
 - `moonstone/internal/Scroller.UiScroller` component, transferred from `ui/Scroller.UiScroller`
+- `moonstone/internal/SlotItem` component, transferred from `ui/SlotItem`
+- `moonstone/internal/ToggleItem` component, transferred from `ui/ToggleItem`
+- `moonstone/internal/VirtualList.UiVirtalListBase` component, transferred from `ui/VirtualList.UiVirtalListBase`
 - `moonstone/VirtualList.VirtualList`, `moonstone/VirtualList.VirtualGridList` prop `data-webos-voice-focused`, `data-webos-voice-disabled`, and `data-webos-voice-group-label`
 
 ### Fixed